deps/v8/test/inspector/debugger/wasm-scope-info-expected.txt

Test retrieving scope information from compiled Liftoff frames

Running test: test
Calling instantiate function.
Waiting for wasm script to be parsed.
Got wasm script!
Setting breakpoint on line 2 (first instruction) of third function
{
    columnNumber : 169
    lineNumber : 0
    scriptId : <scriptId>
}
Paused:
Script wasm://wasm/e33badc2 byte offset 169: Wasm opcode 0x20 (kExprLocalGet)
Scope:
at C (interpreted) (0:169):
 - scope (wasm-expression-stack):
 - scope (local):
   $i32_arg: 42 (number)
   $i32_local: 0 (number)
   $var2: 0 (number)
 - scope (module):
   instance: exports: "exported_global" (Global), "exported_memory" (Memory), "exported_table" (Table), "main" (Function)
   module: Module
   $exported_memory: Memory(1)
   globals: "$exported_global": 0 (number)
at B (liftoff) (0:158):
 - scope (wasm-expression-stack):
   0: 42 (number)
   1: 3 (number)
 - scope (local):
   $i32_arg: 42 (number)
   $i32_local: 0 (number)
   $f32_local: 7.199999809265137 (number)
   $0: 0 (number)
   $var5: 0 (number)
   $v128_local: Uint8Array(16)
 - scope (module):
   instance: exports: "exported_global" (Global), "exported_memory" (Memory), "exported_table" (Table), "main" (Function)
   module: Module
   $exported_memory: Memory(1)
   globals: "$exported_global": 0 (number)
at A (liftoff) (0:128):
 - scope (wasm-expression-stack):
 - scope (local):
   $var0: 42 (number)
 - scope (module):
   instance: exports: "exported_global" (Global), "exported_memory" (Memory), "exported_table" (Table), "main" (Function)
   module: Module
   $exported_memory: Memory(1)
   globals: "$exported_global": 0 (number)
at (anonymous) (0:17):
 - scope (global):
   -- skipped globals

Paused:
Script wasm://wasm/e33badc2 byte offset 171: Wasm opcode 0x24 (kExprGlobalSet)
Scope:
at C (interpreted) (0:171):
 - scope (wasm-expression-stack):
   0: 42 (number)
 - scope (local):
   $i32_arg: 42 (number)
   $i32_local: 0 (number)
   $var2: 0 (number)
 - scope (module):
   instance: exports: "exported_global" (Global), "exported_memory" (Memory), "exported_table" (Table), "main" (Function)
   module: Module
   $exported_memory: Memory(1)
   globals: "$exported_global": 0 (number)
at B (liftoff) (0:158):
 - scope (wasm-expression-stack):
   0: 42 (number)
   1: 3 (number)
 - scope (local):
   $i32_arg: 42 (number)
   $i32_local: 0 (number)
   $f32_local: 7.199999809265137 (number)
   $0: 0 (number)
   $var5: 0 (number)
   $v128_local: Uint8Array(16)
 - scope (module):
   instance: exports: "exported_global" (Global), "exported_memory" (Memory), "exported_table" (Table), "main" (Function)
   module: Module
   $exported_memory: Memory(1)
   globals: "$exported_global": 0 (number)
at A (liftoff) (0:128):
 - scope (wasm-expression-stack):
 - scope (local):
   $var0: 42 (number)
 - scope (module):
   instance: exports: "exported_global" (Global), "exported_memory" (Memory), "exported_table" (Table), "main" (Function)
   module: Module
   $exported_memory: Memory(1)
   globals: "$exported_global": 0 (number)
at (anonymous) (0:17):
 - scope (global):
   -- skipped globals

Paused:
Script wasm://wasm/e33badc2 byte offset 173: Wasm opcode 0x41 (kExprI32Const)
Scope:
at C (interpreted) (0:173):
 - scope (wasm-expression-stack):
 - scope (local):
   $i32_arg: 42 (number)
   $i32_local: 0 (number)
   $var2: 0 (number)
 - scope (module):
   instance: exports: "exported_global" (Global), "exported_memory" (Memory), "exported_table" (Table), "main" (Function)
   module: Module
   $exported_memory: Memory(1)
   globals: "$exported_global": 42 (number)
at B (liftoff) (0:158):
 - scope (wasm-expression-stack):
   0: 42 (number)
   1: 3 (number)
 - scope (local):
   $i32_arg: 42 (number)
   $i32_local: 0 (number)
   $f32_local: 7.199999809265137 (number)
   $0: 0 (number)
   $var5: 0 (number)
   $v128_local: Uint8Array(16)
 - scope (module):
   instance: exports: "exported_global" (Global), "exported_memory" (Memory), "exported_table" (Table), "main" (Function)
   module: Module
   $exported_memory: Memory(1)
   globals: "$exported_global": 42 (number)
at A (liftoff) (0:128):
 - scope (wasm-expression-stack):
 - scope (local):
   $var0: 42 (number)
 - scope (module):
   instance: exports: "exported_global" (Global), "exported_memory" (Memory), "exported_table" (Table), "main" (Function)
   module: Module
   $exported_memory: Memory(1)
   globals: "$exported_global": 42 (number)
at (anonymous) (0:17):
 - scope (global):
   -- skipped globals

Paused:
Script wasm://wasm/e33badc2 byte offset 175: Wasm opcode 0x21 (kExprLocalSet)
Scope:
at C (interpreted) (0:175):
 - scope (wasm-expression-stack):
   0: 47 (number)
 - scope (local):
   $i32_arg: 42 (number)
   $i32_local: 0 (number)
   $var2: 0 (number)
 - scope (module):
   instance: exports: "exported_global" (Global), "exported_memory" (Memory), "exported_table" (Table), "main" (Function)
   module: Module
   $exported_memory: Memory(1)
   globals: "$exported_global": 42 (number)
at B (liftoff) (0:158):
 - scope (wasm-expression-stack):
   0: 42 (number)
   1: 3 (number)
 - scope (local):
   $i32_arg: 42 (number)
   $i32_local: 0 (number)
   $f32_local: 7.199999809265137 (number)
   $0: 0 (number)
   $var5: 0 (number)
   $v128_local: Uint8Array(16)
 - scope (module):
   instance: exports: "exported_global" (Global), "exported_memory" (Memory), "exported_table" (Table), "main" (Function)
   module: Module
   $exported_memory: Memory(1)
   globals: "$exported_global": 42 (number)
at A (liftoff) (0:128):
 - scope (wasm-expression-stack):
 - scope (local):
   $var0: 42 (number)
 - scope (module):
   instance: exports: "exported_global" (Global), "exported_memory" (Memory), "exported_table" (Table), "main" (Function)
   module: Module
   $exported_memory: Memory(1)
   globals: "$exported_global": 42 (number)
at (anonymous) (0:17):
 - scope (global):
   -- skipped globals

Paused:
Script wasm://wasm/e33badc2 byte offset 177: Wasm opcode 0x0b (kExprEnd)
Scope:
at C (interpreted) (0:177):
 - scope (wasm-expression-stack):
 - scope (local):
   $i32_arg: 42 (number)
   $i32_local: 47 (number)
   $var2: 0 (number)
 - scope (module):
   instance: exports: "exported_global" (Global), "exported_memory" (Memory), "exported_table" (Table), "main" (Function)
   module: Module
   $exported_memory: Memory(1)
   globals: "$exported_global": 42 (number)
at B (liftoff) (0:158):
 - scope (wasm-expression-stack):
   0: 42 (number)
   1: 3 (number)
 - scope (local):
   $i32_arg: 42 (number)
   $i32_local: 0 (number)
   $f32_local: 7.199999809265137 (number)
   $0: 0 (number)
   $var5: 0 (number)
   $v128_local: Uint8Array(16)
 - scope (module):
   instance: exports: "exported_global" (Global), "exported_memory" (Memory), "exported_table" (Table), "main" (Function)
   module: Module
   $exported_memory: Memory(1)
   globals: "$exported_global": 42 (number)
at A (liftoff) (0:128):
 - scope (wasm-expression-stack):
 - scope (local):
   $var0: 42 (number)
 - scope (module):
   instance: exports: "exported_global" (Global), "exported_memory" (Memory), "exported_table" (Table), "main" (Function)
   module: Module
   $exported_memory: Memory(1)
   globals: "$exported_global": 42 (number)
at (anonymous) (0:17):
 - scope (global):
   -- skipped globals

Paused:
Script wasm://wasm/e33badc2 byte offset 160: Wasm opcode 0x1a (kExprDrop)
Scope:
at B (liftoff) (0:160):
 - scope (wasm-expression-stack):
   0: 42 (number)
   1: 3 (number)
 - scope (local):
   $i32_arg: 42 (number)
   $i32_local: 0 (number)
   $f32_local: 7.199999809265137 (number)
   $0: 0 (number)
   $var5: 0 (number)
   $v128_local: Uint8Array(16)
 - scope (module):
   instance: exports: "exported_global" (Global), "exported_memory" (Memory), "exported_table" (Table), "main" (Function)
   module: Module
   $exported_memory: Memory(1)
   globals: "$exported_global": 42 (number)
at A (liftoff) (0:128):
 - scope (wasm-expression-stack):
 - scope (local):
   $var0: 42 (number)
 - scope (module):
   instance: exports: "exported_global" (Global), "exported_memory" (Memory), "exported_table" (Table), "main" (Function)
   module: Module
   $exported_memory: Memory(1)
   globals: "$exported_global": 42 (number)
at (anonymous) (0:17):
 - scope (global):
   -- skipped globals

Paused:
Script wasm://wasm/e33badc2 byte offset 161: Wasm opcode 0x1a (kExprDrop)
Scope:
at B (liftoff) (0:161):
 - scope (wasm-expression-stack):
   0: 42 (number)
 - scope (local):
   $i32_arg: 42 (number)
   $i32_local: 0 (number)
   $f32_local: 7.199999809265137 (number)
   $0: 0 (number)
   $var5: 0 (number)
   $v128_local: Uint8Array(16)
 - scope (module):
   instance: exports: "exported_global" (Global), "exported_memory" (Memory), "exported_table" (Table), "main" (Function)
   module: Module
   $exported_memory: Memory(1)
   globals: "$exported_global": 42 (number)
at A (liftoff) (0:128):
 - scope (wasm-expression-stack):
 - scope (local):
   $var0: 42 (number)
 - scope (module):
   instance: exports: "exported_global" (Global), "exported_memory" (Memory), "exported_table" (Table), "main" (Function)
   module: Module
   $exported_memory: Memory(1)
   globals: "$exported_global": 42 (number)
at (anonymous) (0:17):
 - scope (global):
   -- skipped globals

Paused:
Script wasm://wasm/e33badc2 byte offset 162: Wasm opcode 0x0b (kExprEnd)
Scope:
at B (liftoff) (0:162):
 - scope (wasm-expression-stack):
 - scope (local):
   $i32_arg: 42 (number)
   $i32_local: 0 (number)
   $f32_local: 7.199999809265137 (number)
   $0: 0 (number)
   $var5: 0 (number)
   $v128_local: Uint8Array(16)
 - scope (module):
   instance: exports: "exported_global" (Global), "exported_memory" (Memory), "exported_table" (Table), "main" (Function)
   module: Module
   $exported_memory: Memory(1)
   globals: "$exported_global": 42 (number)
at A (liftoff) (0:128):
 - scope (wasm-expression-stack):
 - scope (local):
   $var0: 42 (number)
 - scope (module):
   instance: exports: "exported_global" (Global), "exported_memory" (Memory), "exported_table" (Table), "main" (Function)
   module: Module
   $exported_memory: Memory(1)
   globals: "$exported_global": 42 (number)
at (anonymous) (0:17):
 - scope (global):
   -- skipped globals

Paused:
Script wasm://wasm/e33badc2 byte offset 130: Wasm opcode 0x0b (kExprEnd)
Scope:
at A (liftoff) (0:130):
 - scope (wasm-expression-stack):
 - scope (local):
   $var0: 42 (number)
 - scope (module):
   instance: exports: "exported_global" (Global), "exported_memory" (Memory), "exported_table" (Table), "main" (Function)
   module: Module
   $exported_memory: Memory(1)
   globals: "$exported_global": 42 (number)
at (anonymous) (0:17):
 - scope (global):
   -- skipped globals

Paused:
instance.exports.main(42)

Scope:
at (anonymous) (1:31):
 - scope (global):
   -- skipped globals

exports.main returned. Test finished.
